Organized file structure, fixed creation/edit product errors
