+++ src/main/java/com/JavaMart/Servlets/EditProductServlet.java
@@ -1,5 +1,6 @@
-package com.JavaMart;
+package com.JavaMart.Servlets;
 
+import com.JavaMart.Classes.*;
 import java.io.IOException;
 import java.util.List;
 import javax.servlet.ServletException;
@@ -15,7 +16,8 @@ public class EditProductServlet extends HttpServlet{
 	public void doGet(HttpServletRequest req, HttpServletResponse res) throws IOException, ServletException {
 		String slug = req.getPathInfo();
 		Product product = Product.GetProductBySlug(slug);
-		//System.out.println(product);
+		System.out.println(product.getName());
+		
 		req.setAttribute("product", product);
 		req.setAttribute("slug", slug);
 		req.getRequestDispatcher("/pages/edit_product.jsp").forward(req, res);
