+++ src/main/java/com/JavaMart/Servlets/LogoutServlet.java
@@ -1,4 +1,4 @@
-package com.JavaMart;
+package com.JavaMart.Servlets;
 
 import java.io.IOException;
 import java.util.List;
