+++ src/main/java/com/JavaMart/Servlets/ProductDetailsServlet.java
@@ -1,4 +1,7 @@
-package com.JavaMart;
+package com.JavaMart.Servlets;
+
+import com.JavaMart.Classes.*;
+import com.JavaMart.Classes.User.Staff;
 
 import java.io.IOException;
 import java.util.List;
@@ -14,7 +17,7 @@ public class ProductDetailsServlet extends HttpServlet {
 		
 	public void doGet(HttpServletRequest req, HttpServletResponse res) throws IOException, ServletException {
 		String slug = req.getPathInfo();
-		//System.out.println(slug);
+		System.out.println(slug);
 		Product product = Product.GetProductBySlug(slug);
 		req.setAttribute("product", product);
 		req.getRequestDispatcher("/pages/product_details.jsp").forward(req, res);
@@ -23,6 +26,8 @@ public class ProductDetailsServlet extends HttpServlet {
 	
 	protected void doPost(HttpServletRequest req, HttpServletResponse res) throws ServletException, IOException {
 		String slug = req.getPathInfo();
+		Product product = Product.GetProductBySlug(slug);
+		String oldSKU = product.getSKU();
 		String productName = req.getParameter("newName");
 		String productDesc = req.getParameter("newDesc");
 		String productVendor = req.getParameter("newVendor");
@@ -31,9 +36,9 @@ public class ProductDetailsServlet extends HttpServlet {
 		String productSKU = req.getParameter("newSKU");
 		
 		//System.out.println(productSlug + "<");
-		Staff.UpdateProduct(productSKU, productName, productDesc, productVendor, productSlug, productPrice);
+		Staff.UpdateProduct(oldSKU, productName, productDesc, productVendor, productSlug, productPrice, productSKU);
 		slug = slug.substring(1);
-		res.sendRedirect("./" + slug);
+		res.sendRedirect("./" + productSlug);
 	}
 
 }
