+++ src/main/webapp/pages/create_product.jsp
@@ -1,7 +1,7 @@
 <%@ page language="java" contentType="text/html; charset=ISO-8859-1"
     pageEncoding="ISO-8859-1"%>
 <%@ page import="java.util.List" %>
-<%@ page import="com.JavaMart.Product" %>
+<%@ page import="com.JavaMart.Classes.Product" %>
 <%@ page import="com.JavaMart.ProductFactory" %>
 <!DOCTYPE html>
 <html>
